Fixed: Lipsync Added: More 3D Space audio options

+++ Echokraut/DataClasses/Configuration.cs
@@ -35,7 +35,7 @@ public class Configuration : IPluginConfiguration
     public bool VoiceBubble { get; set; } = false;
     public bool VoiceBubblesInCity { get; set; } = false;
     public bool VoiceSourceCam { get; set; } = false;
-    public int VoiceBubbleAudibleRange { get; set; } = 1;
+    public float VoiceBubbleAudibleRange { get; set; } = 1f;
     public bool VoiceChat { get; set; } = false;
     public bool VoiceChatPlayer { get; set; } = false;
     public bool VoiceChatSay { get; set; } = false;
